% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCheck.R
\name{boxplotScore}
\alias{boxplotScore}
\title{Quality checks for the score, if smoking variable is available}
\usage{
boxplotScore(probScore, questVariable)
}
\arguments{
\item{probScore}{probability score, returned from the smokeScore function, specifying "prob"}

\item{questVariable}{a binary variable for expsure to maternal smoking during pregnancy, available
in the researchers study data}
}
\value{
\code{boxplotScore} as boxplot for score and study information on maternal smoking.
}
\description{
Quality checks for the score, if smoking variable is available
}
\examples{
 boxplotScore(probScore, questVariable)
}
