% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CpGs450K}
\alias{CpGs450K}
\title{CpGs for the score in 450k.}
\format{An object of class \code{character} of length 204.}
\source{
Machine learning score paper Rauschert et al. under review
}
\usage{
CpGs450K
}
\description{
Vector with all CpGs required for the 450k score
}
\keyword{datasets}
