% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CpGsEPIC}
\alias{CpGsEPIC}
\title{CpGs for the score in EPIC.}
\format{An object of class \code{character} of length 181.}
\source{
Machine learning score paper Rauschert et al. under review
}
\usage{
CpGsEPIC
}
\description{
Vector with all CpGs required for the EPIC score
}
\keyword{datasets}
