% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAsmokeR.R
\docType{package}
\name{DNAsmokeR-package}
\alias{DNAsmokeR}
\alias{DNAsmokeR-package}
\title{DNAsmokeR: DNA methylation based score for in utero exposure to maternal smoking}
\description{
This package contains functions to create a whole blood DNA-methylation based
	in utero exposure score to maternal smoking. This score was developed using elastic net regression.
	The weights provided here can be applied to Illumina HumanMethylation450 or EPIC array data. The
	output can both be a score or a binary classification. This packages contains both the function
	necessary for the score calculation and the weights for the CpGs required for the calculation.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Hobbeist/DNAsmokeR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Rauschert \email{Sebastian.Rauschert@telethonkids.org.au}

}
\keyword{internal}
