% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCheck.R
\name{scoreMatrix}
\alias{scoreMatrix}
\title{Quality checks for the score, if smoking variable is available}
\usage{
scoreMatrix(classScore, questVariable)
}
\arguments{
\item{classScore}{returned from the smokeScore function, specifying "class"}

\item{questVariable}{a binary variable for expsure to maternal smoking during pregnancy, available
in the researchers study data}
}
\value{
\code{scoreMatrix} as confusion matrix.
}
\description{
Quality checks for the score, if smoking variable is available
}
\examples{
\dontrun{
 scoreMatrix(classScore, questVariable)
 }
}
