% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySmokeR.R
\name{shinySmokeR}
\alias{shinySmokeR}
\title{Shiny app to make the application of the smoking score as easy as possible}
\usage{
shinySmokeR()
}
\value{
\code{shinySmokeR} to launch shiny app
}
\description{
Shiny app to make the application of the smoking score as easy as possible
}
\examples{
 shinySmokeR()
}
