% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smokeScore.R
\name{smokeScore}
\alias{smokeScore}
\title{Create a DNA methylation based in utero exposure score to maternal
smoking.}
\usage{
smokeScore(data, ARRAY, class)
}
\arguments{
\item{data}{A data frame with CpGs in the columns and study participants in the rows.}

\item{ARRAY}{specifying array type: either "450k" or "EPIC".}

\item{class}{specifying the desired output: either "prob" or "class".}
}
\value{
\code{smokeScore} as probability or classification.
}
\description{
Create a DNA methylation based in utero exposure score to maternal
smoking.
}
\examples{
 smokeScore(data, "450k", "prob")
 smokeScore(data, "EPIC", "class")
}
